# SYMFONY_API_Voluntariado4v/api_voluntariado_4v/openapi.yaml
openapi: 3.0.0
info:
  title: 'API Voluntariado 4V'
  description: 'API REST para la gesti├│n de voluntariado escolar (Alumnos, Coordinadores y ONGs).'
  version: 1.0.0
paths:
...
  /usuarios:
    post:
      tags:
        - Usuarios
      operationId: post_api_crear_usuario
      requestBody:
        description: 'Datos b├ísicos del usuario'
        required: true
        content:
          application/json:
            schema:
              properties:
                correo:
                  type: string
                  example: usuario@test.com
                google_id:
                  type: string
                  example: goog_123abc
                id_rol:
                  type: integer
                  example: 2
              type: object
      responses:
        '201':
          description: 'Usuario creado con ├®xito'
        '409':
          description: 'Conflicto: El correo o GoogleID ya existen'
        '404':
          description: 'Rol no encontrado'
  '/usuarios/{id}':
    delete:
      tags:
        - Usuarios
      operationId: delete_api_borrar_usuario
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Usuario eliminado (Soft Delete aplicado)'
        '404':
          description: 'Usuario no encontrado'
  '/usuarios/{id}/imagen':
    post:
      tags:
        - Usuarios
      operationId: post_api_upload_imagen_usuario
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                imagen:
                  description: 'Archivo de imagen (jpg, jpeg, png, webp). M├íximo 5MB.'
                  type: string
                  format: binary
              type: object
      responses:
        '200':
          description: 'Imagen de perfil actualizada correctamente'
        '400':
          description: 'Error en la validaci├│n del archivo'
        '404':
          description: 'Usuario no encontrado'
        '500':
          description: 'Error de escritura en disco'
  '/usuarios/{id}/rol':
    put:
      tags:
        - Usuarios
      operationId: put_api_cambiar_rol_usuario
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 'Nuevo rol para el usuario'
        required: true
        content:
          application/json:
            schema:
              properties:
                id_rol:
                  description: '1=Coordinador, 2=Organizaci├│n, 3=Voluntario'
                  type: integer
                  example: 2
              type: object
      responses:
        '200':
          description: 'Rol actualizado correctamente'
          content:
            application/json:
              schema:
                properties:
                  mensaje: { type: string }
                  usuario: { type: object }
                  advertencia: { type: string, nullable: true }
                type: object
        '404':
          description: 'Usuario o Rol no encontrado'
        '400':
          description: 'Datos inv├ílidos'
...
components:
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
...